+++ CHANGELOG.md
@@ -1,5 +1,23 @@
 # @tribesxyz/ayaos
 
+## 0.1.34
+
+### Patch Changes
+
+- 0500d7d: flip ayaLogger arg order
+
+## 0.1.33
+
+### Patch Changes
+
+- e546827: export farcaster service
+
+## 0.1.32
+
+### Patch Changes
+
+- 07dfacf: expose farcaster getCast
+
 ## 0.1.31
 
 ### Patch Changes